+++ ofproto/netflow.h
@@ -50,7 +50,7 @@ struct netflow_flow {
     uint64_t byte_count_off;      /* Byte count at last time out. */
 
     ofp_port_t output_iface;      /* Output interface index. */
-    uint8_t tcp_flags;            /* Bitwise-OR of all TCP flags seen. */
+    uint16_t tcp_flags;           /* Bitwise-OR of all TCP flags seen. */
 };
 
 struct netflow *netflow_create(void);
@@ -68,7 +68,7 @@ void netflow_flow_init(struct netflow_flow *);
 void netflow_flow_clear(struct netflow_flow *);
 void netflow_flow_update_time(struct netflow *, struct netflow_flow *,
                               long long int used);
-void netflow_flow_update_flags(struct netflow_flow *, uint8_t tcp_flags);
+void netflow_flow_update_flags(struct netflow_flow *, uint16_t tcp_flags);
 bool netflow_active_timeout_expired(struct netflow *, struct netflow_flow *);
 
 #endif /* netflow.h */
